Expected Value with targeted overtime
Evaluating the model with an expected value framework which helps us get the Return of Investment for various scenarios we take based on the parameters we choose. In this particular code, we look for a case where we calculate Return of Investment for OverTime policy and also target certain employees who work overTime.